app/oauth2.py
from datetime import datetime, timedelta

from jose import jwt, JWTError
from fastapi import HTTPException, Depends, status
from fastapi.security import OAuth2PasswordBearer
from sqlalchemy.orm import Session

from .config import settings
from .schemas import TokenData
from .database import get_db
from .models import User

oauth2_scheme = OAuth2PasswordBearer(tokenUrl='login')


def create_access_token(data: dict):
    to_encode = data.copy()
    expiration_time = datetime.utcnow(
    ) + timedelta(minutes=settings.jwt_token_expire_minutes)
    to_encode.update({'exp': expiration_time})
    encoded_jwt = jwt.encode(
        to_encode, settings.jwt_secret_key, algorithm=settings.jwt_algorithm)
    return encoded_jwt


def verify_access_token(token: str, credentials_exception: HTTPException):
    try:
        payload = jwt.decode(token, settings.jwt_secret_key,
                             algorithms=[settings.jwt_algorithm])
        id: str = payload.get('user_id')
        if id is None:
            raise credentials_exception
        token_data = TokenData(id=id)
    except JWTError:
        raise credentials_exception
    return token_data


def get_current_user(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
    credentials_exception = HTTPException(
        status.HTTP_401_UNAUTHORIZED,
        detail='Invalid credentials',
        headers={'WWW-Authenticate': 'Bearer'}
    )
    token_data = verify_access_token(token, credentials_exception)
    user_on_db = db.query(User).filter(token_data.id == User.id).first()
    if not user_on_db:
        raise credentials_exception
    return user_on_db
